Rename `auth` folders and files
- `hooks/auth/signupClient` -> `hooks/signup` & `hooks/verify`
- `hooks/auth/signupServer` -> `lib/cookies`
